+++ Common/src/main/java/net/potionstudios/biomeswevegone/client/renderer/entity/wreath/WreathRenderer.java
@@ -11,7 +11,6 @@ import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.core.Direction;
 import net.minecraft.world.phys.Vec3;
-import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import net.potionstudios.biomeswevegone.client.model.ModelAccess;
 import net.potionstudios.biomeswevegone.client.renderer.entity.state.WreathRenderState;
 import net.potionstudios.biomeswevegone.world.entity.decoration.Wreath;